+++ src/pureTypeScriptSrc/gmc4/Utils.ts
@@ -1,0 +1,11 @@
+// ユーティリティ
+// - 恥ずかしながらビット演算に不慣れなので、自分でわかりにくい演算コードはユーティリティメソッドとユニットテストにくくり出していく。
+export default class Utils{
+	public static Test(): number{
+		return 123;
+	}
+
+	public static RoundTo4Bit(num: number): number{
+		return num &= ~0xfff0;
+	}
+}
